Fonction dajout colonne refaite en asynchrone

+++ frontend/src/app/component/type-erreur/type-erreur.ts
@@ -1,6 +1,6 @@
 import { CommonModule } from '@angular/common';
-import { Component, EventEmitter, Output, inject, input, output } from '@angular/core';
-import { FormArray, FormGroup, ReactiveFormsModule } from '@angular/forms';
+import { Component, EventEmitter, Output, effect, inject, input, output } from '@angular/core';
+import { AbstractControl, FormArray, FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
@@ -31,11 +31,13 @@ import { DetailProjectService } from '../../service/DetailProjectService';
   styleUrl: './type-erreur.css'
 })
 export class TypeErreur {
-  readonly form = input.required<FormGroup>();
+  form = input.required<FormGroup>();
   readonly formEtape = input<FormGroup>();
   readonly colonne = input<string[]>([]);
   readonly erreurs = input<any[]>([]);
   readonly verifier = input<any>();
+  fb = inject(FormBuilder);
+  colonneObserve = input<Observable<{ name: string, id: string | number }[]>>(new Observable<{ name: string, id: string | number }[]>());
   @Output() addLigne = new EventEmitter<void>();
   @Output() removeLigne = new EventEmitter<number>();
   exist = input<boolean>(false);
@@ -45,14 +47,47 @@ export class TypeErreur {
   findIdEtape = input<(id_operation: string) => string>(() => '');
   addErreur = output<any>();
   annuler = output<any>();
+  updateEtape = output<any>();
   valider = false;
+  rollback = false;
 
   readonly filteredOperations = input.required<Observable<Erreur[]>[]>();
 
-  ngOnInit() {
+  constructor() {
+    effect(() => {
+      const form = this.form();
+      if (!form) return;
+      // group = this.fb.group({
+      //     idErreur: [value.id_type_erreur || -1],
+      //     typeErreur: [value.libelle_erreur || '', Validators.required],
+      //     degre: [value.est_majeur ? 1 : 0, Validators.required],
+      //     coef: [value.coef || 0, Validators.required],
+      //     raccourci: [value.raccourci || '', Validators.required],
+      //   });
+      this.ensureControl(form, 'formErreur', this.fb.array([]));
+      let colonne = ['idErreur', 'typeErreur', 'degre', 'coef', 'raccourci'];
+      const formErreur = form.get('formErreur') as FormArray;
 
+      // On boucle sur chaque FormGroup du FormArray
+      formErreur.controls.forEach((fg: AbstractControl) => {
+        if (fg instanceof FormGroup) {
+          colonne.forEach(col => {
+            if (!fg.get(col)) {
+              // fg.addControl(col, this.fb.control(''));
+              this.ensureControl(fg, col, this.fb.control(''));
+            }
+          });
+        }
+      });
+    }
+    );
+  }
+
+
+  ngOnInit() {
+    // this.initFormGroup();
     // console.log('erreur init x', this.erreurs())
-    const form = this.form();
+    // const form = this.form();
     // console.log(form.controls["formErreur"]);
     // console.log('form3 = ', form);
     // console.log('formErreur control = ', form.controls['formErreur']);
@@ -70,6 +105,20 @@ export class TypeErreur {
     //   )
     // );
     // console.log("Form dans erreur", this.form);
+    this.colonneObserve().subscribe(col => {
+      console.log('colonneObserve', col);
+    })
+  }
+
+  private ensureControl(form: FormGroup, name: string, control: AbstractControl) {
+    if (!form.get(name)) {
+      form.addControl(name, control);
+    }
+  }
+
+
+  initFormGroup() {
+  
   }
 
   lister(): string[] {
@@ -82,7 +131,7 @@ export class TypeErreur {
     }
     if (colonne.length < 4) {
       let reste = 4 - colonne.length
-      console.log(reste);
+      // console.log(reste);
       for (let index = 0; index < reste; index++) {
         colonne.push("");
       }
@@ -98,6 +147,13 @@ export class TypeErreur {
 
 
   get formGroup() {
+    // console.log('formGroup liste', (this.form().controls["formErreur"] as FormArray).value);
+    // (this.form().controls["formErreur"] as FormArray).controls.forEach((fg, index) => {
+    //   fg = fg as FormGroup;
+    //   console.log('formGroup fg', fg, index, fg.value);
+
+    // });
+    // return [];
     return (this.form().controls["formErreur"] as FormArray).controls as FormGroup[];
   }
 
@@ -111,12 +167,30 @@ export class TypeErreur {
 
   async updateValue(event: any) {
     let { id, value, name } = event;
-    let { id_projet, libelle, index , est_majeur} = value;
+    console.log("UpdateValue appelé avec :", event);
+    
+    let { id_projet, libelle, index, est_majeur, raccourci } = value;
+    if (libelle && index !== undefined && index !== null) {
+      let doublageInterdit = this.formGroup.map(fg => fg.get('typeErreur')?.value?.toLowerCase()).some((val, idx) => val === libelle.toLowerCase() && idx !== index);
+      if (doublageInterdit) {
+        this.rollback = true;
+        this.trigerRemove(index);
+        alert("Doublon d'erreur non accepté");
+        return;
+      }
+    }
+
     if (typeof value === 'object') {
       if (value.libelle !== null && value.libelle !== undefined) {
         value = { id_projet, libelle };
       } else if (value.est_majeur !== null && value.est_majeur !== undefined) {
         value = est_majeur;
+        this.valider = true;
+      } else if (value.raccourci !== null && value.raccourci !== undefined) {
+        value = raccourci;
+        this.valider = true;
+      } else if (value.coef !== null && value.coef !== undefined) {
+        value = value.coef;
         this.valider = true;
       }
     } else {
@@ -124,18 +198,18 @@ export class TypeErreur {
     }
     console.log("Debounced event:", event, this.valider);
     console.log("Valeur avant vérification:", this.verify(value.libelle), this.valider);
-    console.log("Valeur ajoutée, ", this.erreurs().find(item => (item.type_erreur ?? '').toLowerCase() === (value.libelle ?? '').toLowerCase() ));
+    console.log("Valeur ajoutée, ", this.erreurs().find(item => (item.type_erreur ?? '').toLowerCase() === (value.libelle ?? '').toLowerCase()));
     if (!this.verify(value.libelle) && this.valider) {
       let idTo = -1;
       let nameTo = 'detail_projet.erreur_suggestion:libelle';
       let valueTo = { 'libelle': value.libelle };
-      this.detailService.updateUnitaire(idTo, valueTo, nameTo).then((res:any) => {
+      this.detailService.updateUnitaire(idTo, valueTo, nameTo).then((res: any) => {
         console.log("Option ajoutée avec succès :", res, this.erreurs());
-        if (res.parametre && res.parametre.length > 0 ) {
+        if (res.parametre && res.parametre.length > 0) {
           let newErreur = { id_erreur: res.parametre[0].id_erreur_suggestion, type_erreur: res.parametre[0].libelle };
           // this.erreurs()?.push(newErreur);
           this.addErreur.emit(newErreur);
-          console.log("Valeur ajoutée, ", this.erreurs().find(item => (item.type_erreur ?? '').toLowerCase() === (value.libelle ?? '').toLowerCase() ));
+          console.log("Valeur ajoutée, ", this.erreurs().find(item => (item.type_erreur ?? '').toLowerCase() === (value.libelle ?? '').toLowerCase()));
           this.updateValue(event);
           this.valider = false;
         }
@@ -145,16 +219,16 @@ export class TypeErreur {
       console.log("Valeur non valide, mise à jour annulée.");
       return;
     } else if (this.verify(value.libelle) && this.valider) {
-      console.log("Valeur existante, ", this.erreurs().find(item => (item.type_erreur ?? '').toLowerCase() === (value.libelle ?? '').toLowerCase() ));
-      
+      console.log("Valeur existante, ", this.erreurs().find(item => (item.type_erreur ?? '').toLowerCase() === (value.libelle ?? '').toLowerCase()));
+
       this.detailService.updateUnitaire(id, value, name)
-      .then((res : any) => {
-        console.log("Update result:", res, this.formGroup[index]);
-        this.formGroup[index].patchValue({idErreur: res.parametre[0].id_type_erreur}, { emitEvent: false })
-      })
-      .catch(err => {
-        console.error("Erreur lors de la mise à jour :", err);
-      })
+        .then((res: any) => {
+          console.log("Update result:", res, this.formGroup[index], index,this.formGroup);
+          this.formGroup[index].patchValue({ idErreur: res.parametre[0].id_type_erreur }, { emitEvent: false })
+        })
+        .catch(err => {
+          console.error("Erreur lors de la mise à jour :", err);
+        })
     } else {
       console.log("Valeur non valide, mise à jour annulée.", value);
     }
@@ -162,21 +236,21 @@ export class TypeErreur {
 
   async updateValueCheck(event: any) {
     let { id, value, name } = event;
-    let id_etape = this.takeIdEtape(value.id_operation);
+    let id_etape = this.takeIdEtape(value.champ.col.id);
     let { validite, id_operation, id_type_erreur, champ } = value;
     let updateValue = { 'id_etape_qualite': id_etape, 'id_type_erreur': id_type_erreur };
     console.log("Debounced event:", event, 'update value ', updateValue, 'name', champ);
     let result: any = null;
     if (validite) {
       id = -1;
-      let checker = this.formArray.at(champ.index).get(champ.col);
+      let checker = this.formArray.at(champ.index).get(champ.col.name);
       let form = this.formArray.at(champ.index).value;
       // if (this.verify(updateValue.))
       this.detailService.updateUnitaire(id, updateValue, name).then(res => {
         console.log("Update result:", res);
-        checker?.setValue(form[champ.col], { emitEvent: false });
+        checker?.setValue(validite, { emitEvent: false });
       }).catch(err => {
-        checker?.setValue(form[champ.col], { emitEvent: false });
+        checker?.setValue(form[champ.col.name], { emitEvent: false });
         alert("Erreur lors de l'ajout d'option");
       });
       console.log("Update result insert:", this.formArray.at(champ.index).value);
@@ -189,7 +263,7 @@ export class TypeErreur {
 
   verify(value: string): boolean {
     if (this.erreurs()) {
-      return this.erreurs().some(item => (item.type_erreur ?? '').toLowerCase() === (value ?? '').toLowerCase() );
+      return this.erreurs().some(item => (item.type_erreur ?? '').toLowerCase() === (value ?? '').toLowerCase());
     }
     return false;
   }
@@ -206,15 +280,20 @@ export class TypeErreur {
 
     let id = this.formArray.at(index).get('idErreur')?.value;
     let name = 'detail_projet.type_erreur';
-    this.detailService.deleteDonne(id, name)
-    .then(res => {
-      if (res) {
-        this.removeLigne.emit(index);
-      }
-    })
-    .catch((err: any) => {
-      alert("Erreur lors de la suppression de l'erreur " + err.message);
-    });
+    if (!this.rollback) {
+      this.detailService.deleteDonne(id, name)
+        .then(res => {
+          if (res) {
+            this.removeLigne.emit(index);
+            this.updateValue
+          }
+        })
+        .catch((err: any) => {
+          alert("Erreur lors de la suppression de l'erreur " + err.message);
+        });
+    } else {
+      this.removeLigne.emit(index);
+    }
   }
 
   displayErreur = (id: string): string => {
@@ -253,3 +332,7 @@ export class TypeErreur {
     }
   }
 }
+function uuidv4(): any {
+  throw new Error('Function not implemented.');
+}
+
